requirements.txt
amply==0.1.6
black==24.1.1
click==8.1.7
cloudpickle==3.0.0
cycler==0.11.0
docutils==0.20.1
dsplot==0.9.0
fonttools==4.48.1
func-timeout==3.0.0
graphviz==0.20.1
gymnasium==0.27.0
gymnasium-notices==0.0.1
importlib-metadata==7.0.0
jax-jumpy==1.0.0
kiwisolver==1.4.5
matplotlib==3.5.1
mypy-extensions==1.0.0
networkx==2.7.1
numpy==1.21.5
packaging==23.2
pathspec==0.12.1
Pillow==10.0.0
platformdirs==4.2.0
prettytable==3.6.0
PuLP==2.4
pydot==2.0.0
pygame==2.1.3.dev8
pygraphviz==1.11
pyparsing==3.1.1
PySimpleGUI==4.60.5
python-dateutil==2.8.2
scipy==1.10.1
Shimmy==0.2.1
six==1.16.0
tomli==2.0.1
typing_extensions==4.9.0
wcwidth==0.2.13
zipp==3.17.0
zope.event==5.0
